third_party/blink/web_tests/http/tests/websocket/close-unref-websocket-expected.txt

Test if WebSocket is closed while handshaking and unreferenced, it should fire close event at most once.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 1 is >= countCloseEvent
PASS successfullyParsed is true

TEST COMPLETE
